pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "guardian-crossword-supernote"
version = "0.1.0"
description = "Downloads Guardian crosswords and uploads them to SuperNote cloud automatically"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "tracpants", email = "github.enslave464@passmail.net"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
    "Topic :: Utilities",
]
keywords = ["crossword", "supernote", "guardian", "pdf", "automation"]
requires-python = ">=3.8"
dependencies = [
    "sncloud==0.2.1",
    "python-dotenv==1.1.1",
    "requests==2.32.4",
]

[project.urls]
Homepage = "https://github.com/tracpants/guardian-crossword-supernote"
Repository = "https://github.com/tracpants/guardian-crossword-supernote"
Issues = "https://github.com/tracpants/guardian-crossword-supernote/issues"

[project.scripts]
guardian-crossword = "guardian_crossword.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"